Move tutorial do history
