+++ src/routes.js
@@ -6,6 +6,7 @@ import Home from './pages/home/home';
 import About from './pages/about/about';
 import Ranking from './pages/ranking/ranking';
 import Tutorial from './pages/tutorial/tutorial';
+import History from './pages/history/history';
 
 const Routes = () => (
   <BrowserRouter>
@@ -13,6 +14,7 @@ const Routes = () => (
       <Route exact path="/" component={Home} />
       <Route path="/about" component={About} />
       <Route path="/ranking" component={Ranking} />
+      <Route path="/historia" component={History} />
       <Route path="/tutorial" component={Tutorial} />
     </Switch>
   </BrowserRouter>
